+++ it/src/test/scala/io/github/thediscprog/simexcaching/HazelcastCachingServiceTest.scala
@@ -1,7 +1,8 @@
-package simex.caching
+package io.github.thediscprog.simexcaching
 
 import cats.effect.IO
 import cats.effect.unsafe.implicits.global
+import io.github.thediscprog.simexcaching.config.HazelcastConfig
 import io.github.thediscprog.simexmessaging.messaging.Datum
 import io.github.thediscprog.simexmessaging.test.SimexTestFixture
 import org.scalatest.OptionValues
@@ -12,7 +13,6 @@ import org.scalatest.time.{Millis, Seconds, Span}
 import org.testcontainers.containers.GenericContainer
 import org.testcontainers.utility.DockerImageName
 import org.typelevel.log4cats.slf4j.Slf4jLogger
-import simex.caching.config.HazelcastConfig
 import io.github.thediscprog.slogic.Xor
 import org.typelevel.log4cats.SelfAwareStructuredLogger
