+++ src/main/scala/io/github/thediscprog/simexcaching/config/HazelcastConfig.scala
@@ -1,4 +1,4 @@
-package simex.caching.config
+package io.github.thediscprog.simexcaching.config
 
 import io.circe.Decoder
 import io.circe.generic.semiauto.deriveDecoder
